Apply feedback and move Finitef64 to primitive.rs
